refactor: lint errors
