+++ mocks/constants/ERC721Collections.json
